+++ src/novelystlib/views/todo_chapter_view.py
@@ -1,4 +1,4 @@
-"""Provide a class for viewing and editing "Todo" scene properties.
+"""Provide a class for viewing and editing "Todo" chapter properties.
 
 Copyright (c) 2022 Peter Triesberger
 For further information see https://github.com/peter88213/novelyst
@@ -12,17 +12,12 @@ from novelystlib.widgets.label_entry import LabelEntry
 from novelystlib.widgets.my_string_var import MyStringVar
 
 
-class TodoSceneView(BasicView):
-    """Class for viewing and editing "Todo" scene properties.
-          
+class TodoChapterView(BasicView):
+    """Class for viewing and editing chapter properties.
+      
     Public methods:
         set_data() -- Update the view with element's data.
         apply_changes() -- Apply changes.   
-
-    If one story arc is assigned to a "Todo" scene, this scene is used 
-    for describing this arc. In this case, there is an extra display:    
-    - The number of normal scenes assigned to this arc.
-    - A button to remove all scene assigments to this arc.
     """
 
     def __init__(self, ui):
@@ -43,30 +38,25 @@ class TodoSceneView(BasicView):
         self._arcsEntry = LabelEntry(self._elementInfoWindow, text=_('Arc reference'), textvariable=self._arcs, lblWidth=22)
         self._arcsEntry.pack(anchor=tk.W, pady=2)
 
-        '''
         # Frame for arc specific widgets.
         self._arcFrame = ttk.Frame(self._elementInfoWindow)
         self._nrScenes = ttk.Label(self._arcFrame)
         self._nrScenes.pack(side=tk.LEFT, pady=2)
         ttk.Button(self._arcFrame, text=_('Remove scene assignments'), command=self._removeArcRef).pack(padx=1, pady=2)
-        '''
-
-        ttk.Separator(self._elementInfoWindow, orient=tk.HORIZONTAL).pack(fill=tk.X)
-
-        # 'Tags' entry.
-        self._tags = MyStringVar()
-        LabelEntry(self._elementInfoWindow, text=_('Tags'), textvariable=self._tags, lblWidth=self._LBL_X).pack(anchor=tk.W, pady=2)
 
     def set_data(self, element):
-        """Update the widgets with element's data.
+        """Update the view with element's data.
         
+        - Hide the info window, if the chapter ist the "trash bin". 
+        - Show/hide the "Do not auto-number" button, depending on the chapter type.       
+        - Configure the "Do not auto-number" button, depending on the chapter level.       
         Extends the superclass constructor.
         """
         super().set_data(element)
 
         # Count the scenes assigned to this arc.
         self._scenesAssigned = []
-        arc = self._element.scnArcs
+        arc = self._element.kwVar['Field_Arc_Definition']
         if arc:
             for scId in self._ui.novel.scenes:
                 if self._ui.novel.scenes[scId].scType == 0:
@@ -78,7 +68,6 @@ class TodoSceneView(BasicView):
         # 'Arc reference' entry.
         self._arcs.set(arc)
 
-        '''
         # Frame for arc specific widgets.
         if len(self._scenesAssigned) > 0:
             self._nrScenes['text'] = f'{_("Number of scenes")}: {len(self._scenesAssigned)}'
@@ -87,14 +76,6 @@ class TodoSceneView(BasicView):
         else:
             if self._arcFrame.winfo_manager():
                 self._arcFrame.pack_forget()
-        '''
-
-        # 'Tags' entry.
-        if self._element.tags is not None:
-            self._tagsStr = list_to_string(self._element.tags)
-        else:
-            self._tagsStr = ''
-        self._tags.set(self._tagsStr)
 
     def apply_changes(self):
         """Apply changes.
@@ -102,22 +83,16 @@ class TodoSceneView(BasicView):
         Extends the superclass method.
         """
         # 'Arc reference' entry.
-        '''
-        newArcs = self._arcs.get()
-        if self._element.scnArcs or newArcs:
-            if self._element.scnArcs != newArcs:
-                self._element.scnArcs = newArcs
+        newArcs = self._arcs.get().replace(';', '')
+        if self._element.kwVar['Field_Arc_Definition'] or newArcs:
+            if self._element.kwVar['Field_Arc_Definition'] != newArcs:
+                for scId in self._element.srtScenes:
+                    self._ui.novel.scenes[scId].scnArcs = newArcs
+                self._element.kwVar['Field_Arc_Definition'] = newArcs
 
                 # Use the arc as scene title suffix.
-                newTitle = f'{self._element.scnArcs} - {self._ui.elementTitle.get()}'
+                newTitle = f'{self._element.kwVar["Field_Arc_Definition"]} - {self._ui.elementTitle.get()}'
                 self._ui.elementTitle.set(newTitle)
-                self._ui.isModified = True
-        '''
-        # 'Tags' entry.
-        newTags = self._tags.get()
-        if self._tagsStr or newTags:
-            if newTags != self._tagsStr:
-                self._element.tags = string_to_list(newTags)
                 self._ui.isModified = True
 
         super().apply_changes()